refactor(039-08): split section notes panel (882→324 lines)
- extract NoteTypeBadge, NoteCard, NoteFilterPanel, AddNoteForm
  into SectionNotesPanel/ subdirectory
- remove // @ts-nocheck and fix all types (enum index assertion,
  explicit NoteCardProps spread)
- mark 039-08 done in sprint plan; update README to 9/14 (64%)
