+++ README.md
@@ -59,7 +59,7 @@
 
 - **Бизнес-модель**: Freemium + подписка (Stars Payment в Telegram)
 - **Рынок**: MusicTech + AI + Creator Economy ($12B+ к 2027)
-- **Текущий фокус**: Sprint 039 — архитектурный рефакторинг: **8/14 задач** (57%), 0 нарушений слоёв, god-компоненты разбиты
+- **Текущий фокус**: Sprint 039 — архитектурный рефакторинг: **9/14 задач** (64%), 0 нарушений слоёв, god-компоненты разбиты
 
 ---
 
@@ -91,7 +91,7 @@ gantt
 |  036   | Рефакторинг слоёв + Type Safety |   ✅   |    100%    |
 |  037   | Infrastructure Hardening        |   ✅   |    100%    |
 |  038   | Design System Unification       |   ✅   |   28/28    |
-|  039   | Архитектурный рефакторинг       |   🟡   | 8/14 (57%) |
+|  039   | Архитектурный рефакторинг       |   🟡   | 9/14 (64%) |
 |  040   | Тесты + Export                  |   ⚪   |     0%     |
 
 <sub>Подробнее: [ROADMAP.md](./ROADMAP.md) · [PROJECT_STATUS.md](./PROJECT_STATUS.md)</sub>
